Moved preprocess_units decorator to the utils module
