+++ cvpack/units/__init__.py
@@ -9,6 +9,5 @@ from .units import (  # noqa: F401
     ScalarQuantity,
     Unit,
     VectorQuantity,
-    preprocess_units,
     value_in_md_units,
 )
